requirements.txt
streamlit
langchain==0.3.0
google-generativeai
numpy
pypdf
python-dotenv
python-pptx
langchain-google-genai
langchain-cohere
langchain-groq
